Refactored Asset tests to use a reusable class for the before and after
