+++ storage/src/androidTest/kotlin/com/waz/zclient/storage/userdatabase/UserDatabaseMigrationTest.kt
@@ -1,0 +1,32 @@
+package com.waz.zclient.storage.userdatabase
+
+import com.waz.zclient.storage.DbSQLiteOpenHelper
+import com.waz.zclient.storage.IntegrationTest
+import com.waz.zclient.storage.MigrationTestHelper
+import com.waz.zclient.storage.db.UserDatabase
+import org.junit.After
+import org.junit.Before
+
+abstract class UserDatabaseMigrationTest(val dbName: String, val version: Int) : IntegrationTest() {
+
+    protected lateinit var testOpenHelper: DbSQLiteOpenHelper
+
+    private val databaseHelper: UserDatabaseHelper by lazy {
+        UserDatabaseHelper()
+    }
+
+    protected lateinit var testHelper: MigrationTestHelper
+
+    @Before
+    fun setUp() {
+        testHelper = MigrationTestHelper(UserDatabase::class.java.canonicalName)
+        testOpenHelper = DbSQLiteOpenHelper(getApplicationContext(),
+            dbName, version)
+        databaseHelper.createDatabase(testOpenHelper)
+    }
+
+    @After
+    fun tearDown() {
+        databaseHelper.clearDatabase(testOpenHelper)
+    }
+}
