+++ storage/src/androidTest/kotlin/com/waz/zclient/storage/userdatabase/assets/AssetTables126to127MigrationTests.kt
@@ -1,39 +1,13 @@
 package com.waz.zclient.storage.userdatabase.assets
 
-import com.waz.zclient.storage.DbSQLiteOpenHelper
-import com.waz.zclient.storage.IntegrationTest
-import com.waz.zclient.storage.MigrationTestHelper
 import com.waz.zclient.storage.db.UserDatabase
 import com.waz.zclient.storage.db.users.migration.USER_DATABASE_MIGRATION_126_TO_127
 import com.waz.zclient.storage.di.StorageModule.getUserDatabase
-import com.waz.zclient.storage.userdatabase.UserDatabaseHelper
+import com.waz.zclient.storage.userdatabase.UserDatabaseMigrationTest
 import kotlinx.coroutines.runBlocking
-import org.junit.After
-import org.junit.Before
 import org.junit.Test
 
-class AssetTables126to127MigrationTests : IntegrationTest() {
-
-    private lateinit var testOpenHelper: DbSQLiteOpenHelper
-
-    private val databaseHelper: UserDatabaseHelper by lazy {
-        UserDatabaseHelper()
-    }
-
-    private lateinit var testHelper: MigrationTestHelper
-
-    @Before
-    fun setUp() {
-        testHelper = MigrationTestHelper(UserDatabase::class.java.canonicalName)
-        testOpenHelper = DbSQLiteOpenHelper(getApplicationContext(),
-            TEST_DB_NAME, 126)
-        databaseHelper.createDatabase(testOpenHelper)
-    }
-
-    @After
-    fun tearDown() {
-        databaseHelper.clearDatabase(testOpenHelper)
-    }
+class AssetTables126to127MigrationTests : UserDatabaseMigrationTest(TEST_DB_NAME, 126) {
 
     @Test
     fun givenV1AssetInsertedIntoAssetsTableVersion126_whenMigratedToVersion127_thenAssertDataIsStillIntact() {
